Bouncy headers code re-organization
