+++ GravatarApp/Extensions/View+extensions.swift
@@ -1,0 +1,25 @@
+import SwiftUI
+
+extension View {
+    func contentHeightReader(_ height: Binding<CGFloat>) -> some View {
+        self.background(
+            GeometryReader { geo in
+                Color.clear.onChange(of: geo.size) { _, value in
+                    height.wrappedValue = value.height
+                }.onAppear {
+                    height.wrappedValue = geo.size.height
+                }
+            }
+        )
+    }
+}
+
+extension View {
+    @ViewBuilder func `if`<Content: View>(_ condition: Bool, transform: (Self) -> Content) -> some View {
+        if condition {
+            transform(self)
+        } else {
+            self
+        }
+    }
+}
