+++ GravatarApp/SharedViews/AnimatedHeaderScrollView/BouncyImageBackgroundHeaderView.swift
@@ -1,0 +1,93 @@
+import SwiftUI
+import GravatarUI
+
+struct BouncyImageBackgroundHeaderView<Content>: View where Content: View {
+    let topSafeArea: CGFloat
+    let imageURL: URL?
+    @Binding var forceRefresh: Bool
+
+    let content: () -> Content
+
+    @State private var contentHeight: CGFloat = 0
+
+    private var viewHeight: CGFloat {
+        return contentHeight + topSafeArea
+    }
+
+    var body: some View {
+        GeometryReader { geo in
+            let offset = geo.frame(in: .global).minY
+            let isBouncing = (offset > 0)
+
+            HeaderAvatarView(imageURL: imageURL, showLoading: false, forceRefresh: $forceRefresh) {
+                EmptyView()
+            }
+            .scaledToFill()
+            .frame(width: geo.size.width ,height: isBouncing ? viewHeight + offset : viewHeight)
+            .clipped()
+            .blur(radius: 26, opaque: true)
+            .offset(y: isBouncing ? -offset : 0)
+            .overlay(content: {
+                Color.black.opacity(0.2)
+                    .frame(height: isBouncing ? viewHeight + offset : viewHeight)
+                    .offset(y: isBouncing ? -offset : 0)
+            })
+            .overlay {
+                VStack {
+                    Spacer()
+                    content()
+                        .if(topSafeArea == 0, transform: { view in
+                            view.padding(.top)
+                        })
+                        .padding(.bottom)
+                        .contentHeightReader($contentHeight)
+                }
+                .frame(height: isBouncing ? viewHeight + offset : viewHeight)
+                .offset(y: isBouncing ? -offset : 0)
+            }
+        }
+        .environment(\.colorScheme, .dark)
+        .frame(height: viewHeight)
+    }
+}
+
+struct HeaderAvatarView<Placeholder>: View where Placeholder: View {
+    let imageURL: URL?
+    let showLoading: Bool
+    @Binding var forceRefresh: Bool
+
+    let placeholderView: () -> Placeholder
+
+    var body: some View {
+        AvatarView(
+            url: imageURL,
+            placeholderView: {
+                placeholderView()
+            },
+            oneTimeForceRefresh: $forceRefresh,
+            loadingView: {
+                showLoading ?
+                AnyView(ProgressView().progressViewStyle(.circular))
+                :
+                AnyView(EmptyView())
+            }
+        )
+    }
+}
+
+#Preview {
+    let imageURL = URL(string: "https://1.gravatar.com/avatar/1?size=256")
+
+    ScrollView {
+        BouncyImageBackgroundHeaderView(
+            topSafeArea: 0,
+            imageURL: imageURL,
+            forceRefresh: .constant(false)
+        ) {
+            VStack {
+                Text("Hello world")
+                Text("Drag me down!")
+            }
+        }
+    }
+}
